+++ src/huc/data.h
@@ -8,32 +8,32 @@
 
 extern struct const_array *const_ptr;
 extern struct const_array const_var[MAX_CONST];
-extern long const_val[MAX_CONST_VALUE];
+extern intptr_t const_val[MAX_CONST_VALUE];
 extern char const_data[MAX_CONST_DATA];
-extern long const_val_start;
-extern long const_val_idx;
-extern long const_data_start;
-extern long const_data_idx;
-extern long const_size;
-extern long const_nb;
+extern intptr_t const_val_start;
+extern intptr_t const_val_idx;
+extern intptr_t const_data_start;
+extern intptr_t const_data_idx;
+extern intptr_t const_size;
+extern intptr_t const_nb;
 
 /* storage words */
 
 extern SYMBOL symtab[];
 extern SYMBOL *glbptr, *rglbptr, *locptr;
-extern long ws[];
-extern long *wsptr;
-extern long swstcase[];
-extern long swstlab[];
-extern long swstp;
+extern intptr_t ws[];
+extern intptr_t *wsptr;
+extern intptr_t swstcase[];
+extern intptr_t swstlab[];
+extern intptr_t swstp;
 extern char litq[];
 extern char litq2[];
-extern long litptr;
+extern intptr_t litptr;
 extern struct macro macq[];
-extern long macptr;
+extern intptr_t macptr;
 extern char line[];
 extern char mline[];
-extern long lptr, mptr;
+extern intptr_t lptr, mptr;
 
 extern TAG_SYMBOL tag_table[NUMTAG];	// start of structure tag table
 extern int tag_table_index;		// ptr to next entry
@@ -45,7 +45,7 @@ extern char asmdefs[];
 
 /* miscellaneous storage */
 
-extern long nxtlab,
+extern intptr_t nxtlab,
 	    litlab,
 	    stkp,
 	    zpstkp,
@@ -65,29 +65,29 @@ extern FILE *input, *input2, *output;
 extern FILE *inclstk[];
 
 extern char inclstk_name[INCLSIZ][FILENAMESIZE];
-extern long inclstk_line[];
+extern intptr_t inclstk_line[];
 extern char fname_copy[FILENAMESIZE];
 extern char user_outfile[FILENAMESIZE];
-extern long line_number;
+extern intptr_t line_number;
 
-extern long inclsp;
+extern intptr_t inclsp;
 extern char fname[];
 
 extern char quote[];
 extern SYMBOL *cptr;
-extern long *iptr;
-extern long fexitlab;
-extern long iflevel, skiplevel;
-extern long errfile;
-extern long sflag;
-extern long cdflag;
-extern long verboseflag;
-extern long startup_incl;
-extern long errs;
+extern intptr_t *iptr;
+extern intptr_t fexitlab;
+extern intptr_t iflevel, skiplevel;
+extern intptr_t errfile;
+extern intptr_t sflag;
+extern intptr_t cdflag;
+extern intptr_t verboseflag;
+extern intptr_t startup_incl;
+extern intptr_t errs;
 
-extern long top_level_stkp;
+extern intptr_t top_level_stkp;
 extern int norecurse;
-extern long locals_ptr;
+extern intptr_t locals_ptr;
 extern char current_fn[];
 
 extern struct type *typedefs;
@@ -104,7 +104,7 @@ extern int enum_type_ptr;
 extern int user_short_enums;
 extern int user_signed_char;
 
-extern long output_globdef;
+extern intptr_t output_globdef;
 extern int have_irq_handler;
 extern int have_sirq_handler;
