+++ src/huc/main.h
@@ -18,12 +18,12 @@
 void FEvers (void);
 void usage (char *exename);
 void parse (void);
-long dodcls (long stclass, TAG_SYMBOL *mtag, int is_struct);
+intptr_t dodcls (intptr_t stclass, TAG_SYMBOL *mtag, int is_struct);
 void dumplits (void);
 void dumpglbs (void);
 void errorsummary (void);
 char extension (char *s);
-long assemble (char *s);
+intptr_t assemble (char *s);
 
 void dotypedef (void);
